[FEATURE] Allow package path to be in vendor directory
This allows the NullPackageManager to assume that when any package name is requested, that package's folder can be found two folders up - e.g. in the `vendor/fluidtypo3/` directory. In other words, this allows any FluidTYPO3 dependency to load files correctly, but not other extensions that are not in the `fluidtypo3` composer namespace.